backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pandas==2.2.3
openpyxl==3.1.5
httpx==0.27.2
python-dotenv==1.0.1
email-validator==2.2.0
slowapi==0.1.9
limits==3.13.0
pydantic-settings==2.5.2
